Genre Already Exists Bug
If a user tries to create a new genre with a name that already exists,
the app was throwing an error since the code was refactored to pass back
the id of the new category. Now the function passes 'None' if the
category already exists and the app redirects to showCatalog instead of
showItems.